test(vanilla): align MSW handlers and shared data for SSR/SSG
