+++ packages/vanilla/src/mocks/handlers.js
@@ -1,66 +1,8 @@
 import { http, HttpResponse } from "msw";
 import items from "./items.json";
+import { filterProducts, getUniqueCategories } from "./shared";
 
 const delay = async () => await new Promise((resolve) => setTimeout(resolve, 200));
-
-// 카테고리 추출 함수
-function getUniqueCategories() {
-  const categories = {};
-
-  items.forEach((item) => {
-    const cat1 = item.category1;
-    const cat2 = item.category2;
-
-    if (!categories[cat1]) categories[cat1] = {};
-    if (cat2 && !categories[cat1][cat2]) categories[cat1][cat2] = {};
-  });
-
-  return categories;
-}
-
-// 상품 검색 및 필터링 함수
-function filterProducts(products, query) {
-  let filtered = [...products];
-
-  // 검색어 필터링
-  if (query.search) {
-    const searchTerm = query.search.toLowerCase();
-    filtered = filtered.filter(
-      (item) => item.title.toLowerCase().includes(searchTerm) || item.brand.toLowerCase().includes(searchTerm),
-    );
-  }
-
-  // 카테고리 필터링
-  if (query.category1) {
-    filtered = filtered.filter((item) => item.category1 === query.category1);
-  }
-  if (query.category2) {
-    filtered = filtered.filter((item) => item.category2 === query.category2);
-  }
-
-  // 정렬
-  if (query.sort) {
-    switch (query.sort) {
-      case "price_asc":
-        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
-        break;
-      case "price_desc":
-        filtered.sort((a, b) => parseInt(b.lprice) - parseInt(a.lprice));
-        break;
-      case "name_asc":
-        filtered.sort((a, b) => a.title.localeCompare(b.title, "ko"));
-        break;
-      case "name_desc":
-        filtered.sort((a, b) => b.title.localeCompare(a.title, "ko"));
-        break;
-      default:
-        // 기본은 가격 낮은 순
-        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
-    }
-  }
-
-  return filtered;
-}
 
 export const handlers = [
   // 상품 목록 API
