+++ packages/vanilla/src/mocks/shared.js
@@ -1,0 +1,62 @@
+import items from "./items.json";
+
+export function getUniqueCategories() {
+  const categories = {};
+
+  items.forEach((item) => {
+    const cat1 = item.category1;
+    const cat2 = item.category2;
+
+    if (!categories[cat1]) categories[cat1] = {};
+    if (cat2 && !categories[cat1][cat2]) categories[cat1][cat2] = {};
+  });
+
+  return categories;
+}
+
+export function filterProducts(allProducts, query) {
+  let filtered = [...allProducts];
+
+  if (query.search) {
+    const searchTerm = query.search.toLowerCase();
+    filtered = filtered.filter(
+      (item) => item.title.toLowerCase().includes(searchTerm) || item.brand.toLowerCase().includes(searchTerm),
+    );
+  }
+
+  if (query.category1) {
+    filtered = filtered.filter((item) => item.category1 === query.category1);
+  }
+  if (query.category2) {
+    filtered = filtered.filter((item) => item.category2 === query.category2);
+  }
+
+  if (query.sort) {
+    switch (query.sort) {
+      case "price_asc":
+        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
+        break;
+      case "price_desc":
+        filtered.sort((a, b) => parseInt(b.lprice) - parseInt(a.lprice));
+        break;
+      case "name_asc":
+        filtered.sort((a, b) => a.title.localeCompare(b.title, "ko"));
+        break;
+      case "name_desc":
+        filtered.sort((a, b) => b.title.localeCompare(a.title, "ko"));
+        break;
+      default:
+        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
+    }
+  }
+
+  return filtered;
+}
+
+export function paginateProducts(products, page = 1, limit = 20) {
+  const startIndex = (page - 1) * limit;
+  const endIndex = startIndex + limit;
+  return products.slice(startIndex, endIndex);
+}
+
+export { items };
